Added separate folders for components
