+++ src/App.js
@@ -1,8 +1,8 @@
 import React from 'react';
 import './App.css';
-import Welcome from './components/welcome';
-import Header from './components/header';
-import Article from './components/article';
+import Welcome from './components/welcome/welcome';
+import Header from './components/header/header';
+import Article from './components/article/article';
 
 function App() {
   return (
